Add `HPA`/`HVPA` handling to `HighAvailabilityConfig` webhook. (#7226)
* Handle special case in HighAvailabilityConfig webhook

This commit adds a special case handling, when the maxReplicas count
of H(V)PA objects is lower than the calculated replica count of the
deployment or statefulset.

* Add H(V)PA handling to HighAvailabilityConfig webhook

* Remove unneeded pointer receiver

* Solve minor nits in etcd component

* Adjust Kube-Apiserver HVPA

* Adjust Kube-Controller-Manager HVPA

* Adjust CoreDNS HPA

* Address review comments

+++ pkg/operation/botanist/component/etcd/etcd.go
@@ -561,8 +561,8 @@ func (e *etcd) Deploy(ctx context.Context) error {
 						Labels: hpaLabels,
 					},
 					Spec: hvpav1alpha1.HpaTemplateSpec{
-						MinReplicas: pointer.Int32(int32(replicas)),
-						MaxReplicas: int32(replicas),
+						MinReplicas: pointer.Int32(replicas),
+						MaxReplicas: replicas,
 						Metrics: []autoscalingv2beta1.MetricSpec{
 							{
 								Type: autoscalingv2beta1.ResourceMetricSourceType,
